chat-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
motor==3.6.0
pydantic==2.10.4
pydantic-settings==2.7.1
httpx==0.28.1
google-genai==0.4.0
